requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
cffi==1.15.1
decorator==5.1.1
django-bootstrap3==22.1
django-bootstrap4==22.2
django-braces==1.15.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-filter==22.1
django-phonenumber-field==7.0.0
django-recaptcha==3.0.0
Django-Verify-Email==2.0.3
Faker==15.0.0
libmagic==1.0
misaka==2.1.1
phonenumbers==8.12.56
Pillow==9.2.0
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
six==1.16.0
soupsieve==2.3.2.post1
validators==0.20.0
wincertstore==0.2
